+++ crates/fuel-core/src/service/adapters/txpool.rs
@@ -38,7 +38,6 @@ use fuel_core_types::{
         UtxoId,
     },
     fuel_types::{
-        BlockHeight,
         ContractId,
         Nonce,
     },
@@ -49,6 +48,7 @@ use fuel_core_types::{
             GossipsubMessageInfo,
             TransactionGossipData,
         },
+        txpool::Result as TxPoolResult,
     },
 };
 use std::sync::Arc;
@@ -141,9 +141,10 @@ impl fuel_core_txpool::ports::TxPoolDb for Database {
     }
 }
 
+#[async_trait::async_trait]
 impl GasPriceProvider for StaticGasPrice {
-    fn gas_price(&self, _block_height: BlockHeight) -> Option<u64> {
-        Some(self.gas_price)
+    async fn last_gas_price(&self) -> TxPoolResult<u64> {
+        Ok(self.gas_price)
     }
 }