+++ crates/types/src/services/txpool.rs
@@ -284,7 +284,9 @@ pub fn from_executor_to_status(
 #[non_exhaustive]
 pub enum Error {
     #[error("Gas price not found for block height {0}")]
-    GasPriceNotFound(BlockHeight),
+    GasPriceNotFound(String),
+    #[error("Failed to calculate max block bytes: {0}")]
+    MaxBlockBytes(String),
     #[error("TxPool required that transaction contains metadata")]
     NoMetadata,
     #[error("TxPool doesn't support this type of transaction.")]